Inicializador de datos de ejemplo para la base de datos
Agrega datos de ejemplo a la base de datos cuando se inicia la aplicación. Corrección de entidades
